// apps/admin/src/api/generated/model/eventDtoOutputStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type EventDtoOutputStatus =
  (typeof EventDtoOutputStatus)[keyof typeof EventDtoOutputStatus];

export const EventDtoOutputStatus = {
  draft: "draft",
  scheduled: "scheduled",
  finished: "finished",
  cancelled: "cancelled",
} as const;